requirements.txt
accelerate==1.2.0
albucore==0.0.23
albumentations==2.0.3
annotated-types==0.7.0
archspec==0.2.3
boltons==23.0.0
Brotli==1.0.9
cachetools==5.3.3
certifi==2024.8.30
cffi==1.15.1
charset-normalizer==2.0.4
click==8.1.7
conda==24.11.0
conda-content-trust==0.2.0
conda-libmamba-solver==23.11.1
conda-package-handling==2.2.0
conda_package_streaming==0.9.0
contourpy==1.2.0
cryptography==41.0.3
cycler==0.12.1
diffusers==0.31.0
distro==1.9.0
easydict==1.13
einops==0.8.0
faiss-cpu==1.7.4
filelock==3.9.0
fire==0.7.0
fonttools==4.46.0
frozendict==2.4.2
fsspec==2023.12.1
ftfy==6.2.0
fvcore==0.1.5.post20221221
grad-cam==1.5.0
huggingface-hub==0.26.5
idna==3.4
imageio==2.34.1
imgaug==0.4.0
importlib_metadata==8.5.0
iopath==0.1.10
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.32
jsonpointer==2.1
kiwisolver==1.4.5
kornia==0.7.0
lazy_loader==0.4
libmambapy==1.5.3
llvmlite==0.44.0
lmdb==1.4.1
MarkupSafe==2.1.3
matplotlib==3.8.2
menuinst==2.0.2
mpi4py==3.1.6
mpmath==1.3.0
networkx==3.0
ninja==1.11.1.2
numba==0.61.0
numpy==1.26.4
nvidia-ml-py==12.535.133
nvitop==1.3.2
openai-clip==1.0.1
opencv-python==4.8.1.78
opencv-python-headless==4.10.0.82
packaging==23.1
pandas==2.2.1
Pillow==9.3.0
pip==23.3
platformdirs==3.10.0
pluggy==1.0.0
portalocker==3.0.0
protobuf==5.26.1
psutil==5.9.8
pycosat==0.6.6
pycparser==2.21
pydantic==2.10.6
pydantic_core==2.27.2
pynndescent==0.5.13
pyOpenSSL==23.2.0
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
pytorch-metric-learning==2.3.0
pytorch-wavelets==1.3.0
pytz==2024.1
PyWavelets==1.8.0
PyYAML==6.0.1
regex==2024.5.15
registry==0.4.2
requests==2.31.0
ruamel.yaml==0.17.21
safetensors==0.4.5
scikit-image==0.23.2
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.2
sentencepiece==0.2.0
setuptools==68.0.0
shapely==2.0.4
simsimd==6.2.1
six==1.16.0
stringzilla==3.11.3
sympy==1.12
tabulate==0.9.0
tensorboardX==2.6.2.2
tensorly==0.8.1
termcolor==2.4.0
threadpoolctl==3.2.0
tifffile==2024.5.22
timm==0.9.12
tokenizers==0.21.0
tomli==2.0.1
torch==2.1.1+cu118
torchaudio==2.1.1+cu118
torchvision==0.16.1+cu118
tqdm==4.65.0
transformers==4.47.0
triton==2.1.0
truststore==0.8.0
ttach==0.0.3
typing_extensions==4.12.2
tzdata==2024.1
umap==0.1.1
umap-learn==0.5.7
urllib3==1.26.18
wcwidth==0.2.13
wheel==0.41.2
yacs==0.1.8
zipp==3.21.0
zstandard==0.19.0
